+++ homeassistant/components/http/const.py
@@ -1,7 +1,8 @@
 """HTTP specific constants."""
 from typing import Final
 
-KEY_AUTHENTICATED: Final = "ha_authenticated"
+from homeassistant.helpers.http import KEY_AUTHENTICATED  # noqa: F401
+
 KEY_HASS: Final = "hass"
 KEY_HASS_USER: Final = "hass_user"
 KEY_HASS_REFRESH_TOKEN_ID: Final = "hass_refresh_token_id"
